+++ src/MODULE.cpp
@@ -60,6 +60,7 @@ RCPP_MODULE(class_CGALmesh) {
     .method("isValidFaceGraph", &CGALmesh::isValidFaceGraph)
     .method("isValidHalfedgeGraph", &CGALmesh::isValidHalfedgeGraph)
     .method("isValidPolygonMesh", &CGALmesh::isValidPolygonMesh)
+    .method("merge", &CGALmesh::merge)
     .method("orientToBoundVolume", &CGALmesh::orientToBoundVolume)
     .method("print", &CGALmesh::print)
     .method("removeSelfIntersections", &CGALmesh::removeSelfIntersections)